backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
mysql-connector-python==8.4.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
